Observable.js: Introduce mechanism against zombie callbacks
Event subscription always happens via a listener. That
way, we get more robustness when a listener gets destroyed
before the observable, and vice versa.

See Observable.js for a more detailed explanation, especially
the Listener and Observable classes.

This is now necessary, because screens are getting removed, and
we really don't want any callbacks to trigger after that.